+++ test/runtests.jl
@@ -21,4 +21,12 @@ using ReferenceTests
         @info "Testing feature selection..."
         include("test_feature_selection.jl")
     end
+    @testset "Pooling" begin
+        @info "Testing pooling..."
+        include("test_pooling.jl")
+    end
+    @testset "Normalization" begin
+        @info "Testing normalization..."
+        include("test_normalization.jl")
+    end
 end